feat: refactor my code and create my own custom hooks
