+++ client/src/Components/Layout/Navbar/index.jsx
@@ -1,6 +1,6 @@
-import React, { useEffect, useRef, useState } from "react";
+import React from "react";
 import { Link } from "react-router-dom";
-import axios from "axios";
+// import axios from "axios";
 import {
   Box,
   Button,
@@ -13,51 +13,19 @@ import {
 import MenuIcon from "@mui/icons-material/Menu";
 
 import "./style.css";
+import useNavBar from "../../../Hooks/useNavBar";
 
 function Navbar() {
-  const [searchGyms, setSearchGyms] = useState([]);
-  const [isPending, setIsPending] = useState(false);
-  const [search, setSearch] = useState("");
-  const [isShowMenu, setIsShowMenu] = useState(false);
-
-  const ref = useRef(null);
-
-  const handleBlur = () => {
-    setIsPending(false);
-    setSearch("");
-  };
-
-  // function to handle Click Outside input search box
-  useEffect(() => {
-    const handleClickOutside = (event) => {
-      if (ref.current && !ref.current.contains(event.target)) {
-        handleBlur();
-      }
-    };
-    document.addEventListener("click", handleClickOutside, true);
-    return () => {
-      document.removeEventListener("click", handleClickOutside, true);
-    };
-  }, []);
-
-  // function to set Is Show Menu or not
-  const toggleDrawer = () => {
-    setIsShowMenu(!isShowMenu);
-  };
-
-  // function to handle Search and send to API
-  const handleSearch = async (event) => {
-    setSearch(event.target.value);
-    setIsPending(true);
-
-    // send request to server then set searchGyms to the response
-    const { data } = await axios({
-      method: "get",
-      url: `/api/v1/gyms/search?q=${event.target.value}`,
-    });
-
-    setSearchGyms(data.slice(Math.max(data.length - 10, 0))); // get the last 10 results
-  };
+  const {
+    searchGyms,
+    isPending,
+    search,
+    isShowMenu,
+    toggleDrawer,
+    handleSearch,
+    ref,
+    handleBlur,
+  } = useNavBar();
 
   const listNavbar = [
     { text: "عن فت هاوس", link: "#OfferForYouSection" },
@@ -127,9 +95,8 @@ function Navbar() {
 
   const renderInputSearch = () => {
     return (
-      <div className="search-container">
+      <div className="search-container" ref={ref}>
         <input
-          ref={ref}
           className="search-input"
           type="search"
           value={search}
